Extract shared BaseEngine from MantisEngine
Add shrimpy/base_engine.py with BaseEngine, the MDAEngine subclass shared
by all microscope engines (MantisEngine today; iSIM and Dragonfly next).
It owns the platform-independent behavior:

- hardware sequencing defaults and registration with mmc.mda
- debug logging of property, ROI, and XY stage changes
- autofocus: reads metadata.autofocus, dispatches to the simulated
  demo-PFS method or to the engine's engage_autofocus() hook, and skips
  the event when autofocus is enabled but does not engage
- protecting the autofocus stage from Z writes
- resetting hardware properties in teardown_sequence
- acquire() and the indexed acquisition naming

Microscope engines subclass it and override only what differs.
engage_autofocus() raises NotImplementedError in the base; MantisEngine
implements it with Nikon PFS. The autofocus routines now return a bool
instead of each mutating _autofocus_success, which the base records in
one place.

MantisEngine keeps the mantis specifics: acquisition timeouts, XY stage
speed modulation, DynaTrack, and Nikon PFS. Its DynaTrack setup now runs
after the parent setup_sequence, so the pixel size and grid-plan FOV
sizes reflect the state the setup event leaves the hardware in.

Shared engine tests move to tests/test_base_engine.py, with added
coverage for _set_event_properties, _get_autofocus_z_position, the
logging callbacks, and the NotImplementedError path.
