+++ shrimpy/tests/test_mantis_integration.py
@@ -17,8 +17,9 @@ from iohub import open_ome_zarr
 from iohub.ngff import Plate
 from useq import MDASequence, Position
 
+from shrimpy.base_engine import DEMO_PFS_METHOD
 from shrimpy.config import load_config
-from shrimpy.mantis.mantis_engine import DEMO_PFS_METHOD, MantisEngine
+from shrimpy.mantis.mantis_engine import MantisEngine
 
 # Local copy of the demo MDA config, kept in tests/artifacts so test inputs
 # are independent of the project's runtime configuration files.
